Fix implicit widening issues in multiplications
This CL fixes the violations found by enabling
`bugprone-implicit-widening-of-multiplication-result` in clang-tidy
and enables the check in .clang-tidy.

Bug: 515563723
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/324416
Commit-Queue: Ryan Harrison <rharrison@chromium.org>

+++ src/dawn/common/SlabAllocator.cpp
@@ -78,7 +78,8 @@ SlabAllocatorImpl::SlabAllocatorImpl(Index blocksPerSlab,
       mIndexLinkNodeOffset(Align(objectSize, alignof(IndexLinkNode))),
       mBlockStride(Align(mIndexLinkNodeOffset + u32_sizeof<IndexLinkNode>, objectAlignment)),
       mBlocksPerSlab(blocksPerSlab),
-      mTotalAllocationSize(static_cast<size_t>(mSlabBlocksOffset) + mBlocksPerSlab * mBlockStride) {
+      mTotalAllocationSize(static_cast<size_t>(mSlabBlocksOffset) +
+                           static_cast<size_t>(mBlocksPerSlab) * mBlockStride) {
     DAWN_ASSERT(blocksPerSlab > 0);
     DAWN_ASSERT(IsPowerOfTwo(mAllocationAlignment));
 }
